ToSql/Tls/Collation refactor
- Move ToSql to a module with its own name
- Make it public
- Move Tls to the Client
- Merge two collation modules
