+++ src/lib.rs
@@ -2,22 +2,21 @@
 #![recursion_limit = "512"]
 
 mod client;
-mod prepared;
 mod sql_read_bytes;
+mod to_sql;
 
-mod collation;
 mod error;
 mod macros;
 mod result;
 mod row;
 mod tds;
-mod tls;
 
 pub use client::{AuthMethod, Client, ClientBuilder};
 pub use error::Error;
 pub use result::*;
 pub use row::{Column, Row};
 pub use tds::EncryptionLevel;
+pub use to_sql::ToSql;
 pub use uuid::Uuid;
 
 use sql_read_bytes::*;
