+++ src/macros.rs
@@ -40,14 +40,14 @@ macro_rules! uint_enum {
 macro_rules! to_sql {
     ($target:ident, $( $ty:ty: ($name:expr, $val:expr) ;)* ) => {
         $(
-            impl crate::prepared::ToSql for $ty {
+            impl crate::ToSql for $ty {
                 fn to_sql(&self) -> (&'static str, crate::tds::codec::ColumnData) {
                     let $target = self;
                     ($name, $val)
                 }
             }
 
-            impl crate::prepared::ToSql for Option<$ty> {
+            impl crate::ToSql for Option<$ty> {
                 fn to_sql(&self) -> (&'static str, crate::tds::codec::ColumnData) {
                     let val = match self {
                         None => crate::tds::codec::ColumnData::None,
